fix(FolderSearch): Recursively finds all folders
This changes the behaviour from 0.4.0 that was a bit counter-intuitive.
Now each folder name is searched in the whole hierarchy.

All path in the hierarch matching a name in the settings will be scanned
for `.feature` files. It allows the `Specs/` folder to be placed in any
subdirectory of the project, akin to how `Resources/` work in Unity.